Move specific pytest fixtures back to test_model_output.py
These fixtures have moved around a bit. But now that it looks like
all tests from the ModelOutputHandler class will be in a single file,
let's move the relevant pytest fixtures back to that file for
better clarity.
